+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 with respect to its command line interface and HTTP interface.
 
 ## [Unreleased](//github.com/opentable/sous/compare/0.5.57...HEAD)
+
+### Added
 * Server: HTTP client requests to Singularity log to Kafka
 * Server: HTTP client requests to the Docker registry log to Kafka
 * Server: HTTP server responses log to Kafka
@@ -16,7 +18,6 @@ with respect to its command line interface and HTTP interface.
 ### Fixed
 * CLI + Server: No longer panics when printing a nil Deployment.
 
-### Fixed
 * Server: Profiling configuration - accepts SOUS_PROFILING environment variable correctly.
 
 ### Changed